Buy track cursor follows mouse
There is a square that shows where you can build track and follows the cursor. It follows a grid and was ruddy difficult to program - the square had to be within boundaries, and it had to be deleted if the cursor moved, but the boxes on the top and bottom of the boundary would not get deleted until I did some major tweaking.