redo boundary and simplify track laying
Now there is room for the platforms to reside, which involved only minor changes to the if statement. I used the existing platform as a reference (defo knew it would be useful)

I also tried to make the track rotate automatically based on what tracks were near it, but it ended up failing quite badly. It was just too complicated and it didn't work, so I made all track horizontal. Points will be used to change altitude. Here is the final code before deletion.