Buildentry kind of working
Randomly discovered negative indexing in python - had to Google it to find out why -4 was a valid index. Changing the subtraction of storeCoordy from -15 to -14 helped, but I need to limit the area that can be covered by the cursor. Right now there are double the number of entry points than there should be.