Buildentry complete | tracks complete
I moved the position line up, but now the storeCoordy is not updating. The weeks issue has been fixed though. You won't believe it. I tried subtracting 280, 300, 320, 340 from the definition of storeCoordy, but it turns out all I needed to do was correct an earlier mistake which was that I was using positionto define it, not positionCoord! So annoying!!!