buildEntry complete again!
It didn't work. I have givern waay too many attributes to my new subclass, but it was only expecting 4. Maybe I should copy the parent attributes over... Then there were a few syntax errors in my class code. There was also a bit of an arithmatic error in that bit, and I needed to change the position of the buttons because they were waay off. However. It is a little trippy, but it does now work! Brilliant!