Main game screen tracks almost done
It didn't work because I mis-named the file in the code. One that was fixed, the x-coordinate system worked, but everything was on the top line - not good, but better than expected. It turns out I had forgotten to multiply the y value of the trackLayout array by 40 to make it a coordinate, and just added 140. The points don't work though, bu I just need to extend them. Also I need to make add 280 not 140 to the i value.