main screen tracks and entries complete
Not as hard as I though, just some much simpler arithmetic. First test nothing showed up, but that was because I hadn't bought any of them. When I did, only the left hand side ones showed up. This was because the for loop ranges were not inclusive - so I had to change them from 3 to 4, and 1 to 2. Now it works perfectly!