buy Platform thingy
just a quick colour change of the track and we are good for that. I just need to find out how to buy a track, because it is not detecting a track next to it. Maybe something wrong with the range its looking in. Nope, nothing wrong, its working only when you buy it, but its not saving the purchase. Added a thing to save it but its not doing anything. Fixed it! I basically copied the saving function from the function above which was the entryPoints function, but of course that doesn't save anything to the tracksPlatforms file, so I had to change those and it is working perfectly!