Started Contracts
Haven't considered the R&D section but we will return to that later. For now I need to make a new font for the Contracts section and make some TOCs from the Design section of the Write-up. Once a new font is made, I can resize the button to be more better. There are 8 TOCs planned.