Some more progress
I'm starting to lose it. Just have to get through this. I made a variable that should store the value of the train number, but I need to make it global. God I hate myself for not doing this earlier. I wish I was knew I was allowed to code earlier and that pygame was available in school. Eh, Imma take a break now. Mocks tomorrow, good luck!