Started Timetable
Haven't finished with the contracts, but I think it may be easier to come back to those once I've finished the timetable. I'm thinking I should probably make it so that each contract has a weighting which is based on the number of contracts unlocked from researches. We'll see though. We will see. I think contracts should be paid for but we'll see if we have time for that