clicky functionality added
Yay it works perfectly, except I've just realised that remainingTrains will be reset everytime we exit from the timetable screen. We will need to fix this now, and I have an idea to basically scan for trains in the timetable, and remove one train from remainingTrains if ever we see one. Yeah that should work.