bugfixed contracts
Just need to finish contracts, bugfix R&D, then work on the gameplay itself, and I'll actually be done! This is doable!

Anyway, I need to make the contracts unlockable, so I will need to go back to research and change the result so that there is a new variable that somehow stores how many contracts are unlocked or maybe it stores which contracts are unlocked (it also has to be saved to a file - most if not all the global variables do. Then tutorial, but first, R&D unlocking thingy.