engine almost there
All I have to do is make the incidents (which I have kind of already started) and then I can work on actually timetabling it!

Also If I missed it last time, I had an issue where I got the x and y things on a 2D list mixed up because they are backwards. so yeah.